refactor: UI 문자열 상수 적용
